Remove ip helper functions from C Address module
