+++ src/main/build.rs
@@ -198,7 +198,8 @@ fn run_bindgen(build_common: &ShadowBuildCommon) {
         .allowlist_function("runConfigHandlers")
         .allowlist_function("rustlogger_new")
         .allowlist_function("dns_.*")
-        .allowlist_function("address_.*")
+        .allowlist_function("address_getID")
+        .allowlist_function("address_toNetworkIP")
         .allowlist_function("compatsocket_.*")
         .allowlist_function("workerpool_updateMinHostRunahead")
         .allowlist_function("process_.*")
