+++ src/main/host/descriptor/socket.c
@@ -212,7 +212,7 @@ void legacysocket_setPeerName(LegacySocket* socket, in_addr_t ip, in_port_t port
     if(socket->peerString) {
         g_free(socket->peerString);
     }
-    gchar* ipString = address_ipToNewString(ip);
+    gchar* ipString = util_ipToNewString(ip);
     GString* stringBuffer = g_string_new(ipString);
     g_free(ipString);
     g_string_append_printf(stringBuffer, ":%u", ntohs(port));
@@ -253,7 +253,7 @@ void legacysocket_setSocketName(LegacySocket* socket, in_addr_t ip, in_port_t po
         g_free(socket->boundString);
     }
 
-    gchar* ipString = address_ipToNewString(ip);
+    gchar* ipString = util_ipToNewString(ip);
     GString* stringBuffer = g_string_new(ipString);
     g_free(ipString);
     g_string_append_printf(stringBuffer, ":%u (descriptor %p)", ntohs(port), &socket->super);
